refactor: convert component composable to store to retain cached components
- now: componentStore -> componentEditorStore, useStudioComponents -> componentStore
