+++ frontend/src/stores/componentStore.ts
@@ -1,115 +1,34 @@
 import { defineStore } from "pinia"
-import { ref } from "vue"
-import { studioComponents } from "@/data/studioComponents"
-import { getBlockObject, getBlockInstance, confirm } from "@/utils/helpers"
-import getBlockTemplate from "@/utils/blockTemplate"
+import { markRaw, reactive } from "vue"
+import { createDocumentResource } from "frappe-ui"
 import Block from "@/utils/block"
-import useCanvasStore from "@/stores/canvasStore"
-import { toast } from "vue-sonner"
-import type { StudioComponent, ComponentInput } from "@/types/Studio/StudioComponent"
-import { useStudioComponents } from "@/utils/useStudioComponents"
+import type { StudioComponent } from "@/types/Studio/StudioComponent"
+import { getBlockInstance } from "@/utils/helpers"
+import getBlockTemplate from "@/utils/blockTemplate"
+import { ComponentProps } from "@/types"
 
-const useComponentStore = defineStore("componentStore", () => {
-	const selectedComponent = ref<string | null>(null)
-	const componentInputs = ref<ComponentInput[]>([])
-	const { getComponent, getComponentDoc, cacheComponent, removeCachedComponent } = useStudioComponents()
+export const useComponentStore = defineStore("componentStore", () => {
+	const componentMap = reactive<Map<string, Block>>(new Map())
+	const componentDocMap = reactive<Map<string, StudioComponent>>(new Map())
+	const fetchingComponent = reactive<Set<string>>(new Set())
 
-	async function createComponent(componentName: string, blocks?: Block | null) {
-		const component: any = { component_name: componentName }
-		if (blocks) {
-			component.blocks = getBlockObject(blocks)
-		}
-
-		return studioComponents.insert.submit(
-			component,
-			{
-				onSuccess(data: any) {
-					cacheComponent(data)
-					toast.success("Component created successfully")
-					return data
-				},
-				onError(error: any) {
-					toast.error("Failed to create component", {
-						description: error?.messages?.join(", "),
-					})
-				},
-			},
-		)
-	}
-
-	function saveComponent(blocks: Block, componentName: string) {
-		const payload: any = {
+	async function fetchComponent(componentName: string) {
+		const componentDoc = await createDocumentResource({
+			doctype: "Studio Component",
 			name: componentName,
-			blocks: getBlockObject(blocks),
-		}
-
-		payload.inputs = componentInputs.value.map(input => ({
-			input_name: input.name,
-			type: input.type,
-			description: input.description || "",
-			default_value: input.defaultValue || "",
-			required: 0
-		}))
-
-		studioComponents.setValue.submit(
-			payload,
-			{
-				onSuccess(data: StudioComponent) {
-					cacheComponent(data)
-					toast.success("Component saved successfully")
-				},
-				onError(error: any) {
-					toast.error("Failed to save component", {
-						description: error.messages.join(", "),
-					})
-				},
-			},
-		)
+			auto: true,
+		})
+		await componentDoc.get.promise
+		return componentDoc.doc as StudioComponent
 	}
 
-	async function editComponent(componentId: string) {
-		const componentBlocks = await getComponent(componentId)
-		const componentDoc = getComponentDoc(componentId)
-		const blocks = componentBlocks || getBlockInstance(getBlockTemplate("empty-component"))
-
-		// Load existing inputs from the component doc
-		if (componentDoc && componentDoc.inputs) {
-			componentInputs.value = componentDoc.inputs.map((input: any) => ({
-				name: input.input_name,
-				type: input.type,
-				description: input.description,
-				defaultValue: input.default_value
-			}))
-		} else {
-			componentInputs.value = []
-		}
-
-		const canvasStore = useCanvasStore()
-		canvasStore.editOnCanvas(
-			blocks,
-			(editedBlock) => saveComponent(editedBlock, componentDoc.component_id),
-			"Save Component",
-			componentDoc.component_name,
-			componentDoc.component_id,
-			"component",
-		)
+	async function getComponent(componentName: string): Promise<Block | undefined> {
+		await loadComponent(componentName)
+		return componentMap.get(componentName)
 	}
 
-	async function deleteComponent(component: StudioComponent) {
-		const confirmed = await confirm(
-			`Are you sure you want to delete the component '${component.component_name}'?`,
-		)
-		if (confirmed) {
-			studioComponents.delete
-				.submit(component.component_id)
-				.then(() => {
-					toast.success(`Component '${component.component_name}' deleted successfully`)
-					removeCachedComponent(component.component_id)
-				})
-				.catch(() => {
-					toast.error(`Failed to delete component '${component.component_name}'`)
-				})
-		}
+	function getComponentDoc(componentName: string) {
+		return componentDocMap.get(componentName) as StudioComponent
 	}
 
 	function getComponentName(componentId: string) {
@@ -120,38 +39,68 @@ const useComponentStore = defineStore("componentStore", () => {
 		return componentDoc.component_name
 	}
 
-	function addComponentInput(input: ComponentInput) {
-		componentInputs.value.push(input)
-	}
+	async function loadComponent(componentName: string) {
+		if (!componentMap.has(componentName)) {
+			fetchingComponent.add(componentName);
 
-	function updateComponentInput(index: number, input: ComponentInput) {
-		if (index >= 0 && index < componentInputs.value.length) {
-			componentInputs.value[index] = input
+			try {
+				const componentDoc = await fetchComponent(componentName);
+				cacheComponent(componentDoc)
+			} catch {
+				const missingComponentDoc = {
+					name: componentName,
+					component_id: componentName,
+					component_name: componentName,
+					blocks: JSON.stringify(getBlockTemplate("missing-component")),
+					creation: "",
+					modified: "",
+				};
+				cacheComponent(missingComponentDoc)
+			} finally {
+				fetchingComponent.delete(componentName)
+			}
 		}
 	}
 
-	function removeComponentInput(index: number) {
-		if (index >= 0 && index < componentInputs.value.length) {
-			componentInputs.value.splice(index, 1)
+	function cacheComponent(componentDoc: StudioComponent) {
+		componentDocMap.set(componentDoc.component_id, componentDoc)
+		if (componentDoc.blocks) {
+			componentMap.set(componentDoc.component_id, markRaw(getBlockInstance(componentDoc.blocks)))
 		}
 	}
 
-	function clearComponentInputs() {
-		componentInputs.value = []
+	function removeCachedComponent(componentName: string) {
+		componentMap.delete(componentName)
+		componentDocMap.delete(componentName)
+	}
+
+	function getStudioComponentProps(componentName: string): ComponentProps {
+		const componentDoc = componentDocMap.get(componentName)
+		if (!componentDoc || !componentDoc.inputs) return {}
+
+		const props: ComponentProps = {}
+		componentDoc.inputs.forEach(input => {
+			props[input.input_name] = {
+				type: input.type,
+				default: input.default_value || undefined,
+				inputType: input.type,
+				required: !!input.required,
+				options: input.type === "Select" && Array.isArray(input.default_value) ? input.default_value.map((opt: string) => ({ value: opt, label: opt })) : undefined
+			}
+		})
+		return props
 	}
 
 	return {
-		selectedComponent,
-		componentInputs,
-		createComponent,
-		editComponent,
-		deleteComponent,
+		componentMap,
+		componentDocMap,
+		loadComponent,
+		getComponent,
+		getComponentDoc,
 		getComponentName,
-		// inputs
-		addComponentInput,
-		updateComponentInput,
-		removeComponentInput,
-		clearComponentInputs,
+		cacheComponent,
+		removeCachedComponent,
+		getStudioComponentProps,
 	}
 })
